Move line list convenience functions to a mixin
