+++ jdaviz/configs/cubeviz/helper.py
@@ -1,41 +1,7 @@
 from jdaviz.core.helpers import ConfigHelper
-from jdaviz.core.events import AddLineListMessage
+from ..default.plugins.line_lists.line_list_mixin import LineListMixin
 
-class CubeViz(ConfigHelper):
+class CubeViz(ConfigHelper, LineListMixin):
     """CubeViz Helper class"""
 
     _default_configuration = 'cubeviz'
-
-    ## Helper functions and properties related to line lists
-    def load_line_list(self, line_table, replace=False):
-        """
-        Convenience function to get to the viewer function. Also
-        broadcasts a message so the line list plugin UI can display lines
-        loaded via the notebook.
-        """
-        lt = self.app.get_viewer('spectrum-viewer').load_line_list(line_table,
-                                                                   replace=replace,
-                                                                   return_table=True)
-        add_line_list_message = AddLineListMessage(table=lt, sender=self)
-        self.app.hub.broadcast(add_line_list_message)
-
-
-    def erase_spectral_lines(self, name=None):
-        """Convenience function to get to the viewer function"""
-        self.app.get_viewer('spectrum-viewer').erase_spectral_lines(name=name)
-
-    def plot_spectral_line(self, line):
-        """Convenience function to get to the viewer function"""
-        self.app.get_viewer('spectrum-viewer').plot_spectral_line(line)
-
-    def plot_spectral_lines(self):
-        """Convenience function to get to the viewer function"""
-        self.app.get_viewer('spectrum-viewer').plot_spectral_lines()
-
-    @property
-    def spectral_lines(self):
-        return self.app.get_viewer('spectrum-viewer').spectral_lines
-
-    @property
-    def available_linelists(self):
-        return self.app.get_viewer('spectrum-viewer').available_linelists()
